Move distutils.environment.py to distutils.fcompiler
